+++ clarity/src/vm/analysis/type_checker/mod.rs
@@ -52,7 +52,8 @@ impl FunctionType {
             | StacksEpochId::Epoch23
             | StacksEpochId::Epoch24
             | StacksEpochId::Epoch25
-            | StacksEpochId::Epoch30 => self.check_args_2_1(accounting, args, clarity_version),
+            | StacksEpochId::Epoch30
+            | StacksEpochId::Epoch31 => self.check_args_2_1(accounting, args, clarity_version),
             StacksEpochId::Epoch10 => {
                 return Err(CheckErrors::Expects("Epoch10 is not supported".into()).into())
             }
@@ -75,7 +76,8 @@ impl FunctionType {
             | StacksEpochId::Epoch23
             | StacksEpochId::Epoch24
             | StacksEpochId::Epoch25
-            | StacksEpochId::Epoch30 => {
+            | StacksEpochId::Epoch30
+            | StacksEpochId::Epoch31 => {
                 self.check_args_by_allowing_trait_cast_2_1(db, clarity_version, func_args)
             }
             StacksEpochId::Epoch10 => {