+++ stackslib/src/burnchains/tests/mod.rs
@@ -627,6 +627,7 @@ impl TestBurnchainBlock {
 
         let new_snapshot = sortition_db_handle
             .process_block_txs(
+                false,
                 &parent_snapshot,
                 &header,
                 burnchain,