+++ stackslib/src/chainstate/stacks/db/transactions.rs
@@ -8725,6 +8725,7 @@ pub mod test {
                     StacksEpochId::Epoch24 => self.get_stacks_epoch(5),
                     StacksEpochId::Epoch25 => self.get_stacks_epoch(6),
                     StacksEpochId::Epoch30 => self.get_stacks_epoch(7),
+                    StacksEpochId::Epoch31 => self.get_stacks_epoch(8),
                 }
             }
             fn get_pox_payout_addrs(