+++ stackslib/src/chainstate/stacks/index/test/storage.rs
@@ -164,7 +164,7 @@ fn load_store_trie_m_n_same(m: u64, n: u64, same: bool) {
         ];
         path_bytes[24..32].copy_from_slice(&i.to_be_bytes());
 
-        let path = TriePath::from_bytes(&path_bytes).unwrap();
+        let path = TrieHash::from_bytes(&path_bytes).unwrap();
         let value = TrieLeaf::new(&vec![], &[i as u8; 40].to_vec());
         confirmed_marf.insert_raw(path.clone(), value).unwrap();
     }
@@ -213,7 +213,7 @@ fn load_store_trie_m_n_same(m: u64, n: u64, same: bool) {
             ];
             path_bytes[24..32].copy_from_slice(&i.to_be_bytes());
 
-            let path = TriePath::from_bytes(&path_bytes).unwrap();
+            let path = TrieHash::from_bytes(&path_bytes).unwrap();
 
             // NOTE: may have been overwritten; just check for presence
             assert!(
@@ -235,7 +235,7 @@ fn load_store_trie_m_n_same(m: u64, n: u64, same: bool) {
                 path_bytes[16..24].copy_from_slice(&j.to_be_bytes());
             }
 
-            let path = TriePath::from_bytes(&path_bytes).unwrap();
+            let path = TrieHash::from_bytes(&path_bytes).unwrap();
             let value = TrieLeaf::new(&vec![], &[(i + 128) as u8; 40].to_vec());
 
             new_inserted.push((path.clone(), value.clone()));