+++ stackslib/src/chainstate/stacks/index/test/trie.rs
@@ -137,7 +137,7 @@ fn trie_cursor_try_attach_leaf() {
                 path[i] = 32;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
                 // end of path -- cursor points to the insertion point.
@@ -164,7 +164,7 @@ fn trie_cursor_try_attach_leaf() {
                 let leaf_opt_res = MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap(),
+                    &TrieHash::from_bytes(&path[..]).unwrap(),
                 );
                 assert!(leaf_opt_res.is_ok());
 
@@ -194,7 +194,7 @@ fn trie_cursor_try_attach_leaf() {
                 let leaf_opt_res = MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap(),
+                    &TrieHash::from_bytes(&path[..]).unwrap(),
                 );
                 assert!(leaf_opt_res.is_ok());
 
@@ -250,7 +250,7 @@ fn trie_cursor_promote_leaf_to_node4() {
 
         // add a single leaf
         let mut c = TrieCursor::new(
-            &TriePath::from_bytes(&[
+            &TrieHash::from_bytes(&[
                 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22,
                 23, 24, 25, 26, 27, 28, 29, 30, 31,
             ])
@@ -275,7 +275,7 @@ fn trie_cursor_promote_leaf_to_node4() {
             MARF::get_path(
                 &mut f,
                 &block_header,
-                &TriePath::from_bytes(&[
+                &TrieHash::from_bytes(&[
                     0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
                     22, 23, 24, 25, 26, 27, 28, 29, 30, 31
                 ])
@@ -317,7 +317,7 @@ fn trie_cursor_promote_leaf_to_node4() {
             path[i] = 32;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
             // end of path -- cursor points to the insertion point
@@ -342,7 +342,7 @@ fn trie_cursor_promote_leaf_to_node4() {
             let leaf_opt_res = MARF::get_path(
                 &mut f,
                 &block_header,
-                &TriePath::from_bytes(&path[..]).unwrap(),
+                &TrieHash::from_bytes(&path[..]).unwrap(),
             );
             assert!(leaf_opt_res.is_ok());
 
@@ -372,7 +372,7 @@ fn trie_cursor_promote_leaf_to_node4() {
             let leaf_opt_res = MARF::get_path(
                 &mut f,
                 &block_header,
-                &TriePath::from_bytes(&path[..]).unwrap(),
+                &TrieHash::from_bytes(&path[..]).unwrap(),
             );
             assert!(leaf_opt_res.is_ok());
 
@@ -467,7 +467,7 @@ fn trie_cursor_promote_node4_to_node16() {
                 path[k] = j + 32;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
                 f.open_block(&block_header).unwrap();
@@ -486,7 +486,7 @@ fn trie_cursor_promote_node4_to_node16() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -515,7 +515,7 @@ fn trie_cursor_promote_node4_to_node16() {
             path[k] = 128;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -536,7 +536,7 @@ fn trie_cursor_promote_node4_to_node16() {
                 MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap()
+                    &TrieHash::from_bytes(&path[..]).unwrap()
                 )
                 .unwrap()
                 .unwrap(),
@@ -627,7 +627,7 @@ fn trie_cursor_promote_node16_to_node48() {
                 path[k] = j + 32;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -648,7 +648,7 @@ fn trie_cursor_promote_node16_to_node48() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -677,7 +677,7 @@ fn trie_cursor_promote_node16_to_node48() {
             path[k] = 128;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -698,7 +698,7 @@ fn trie_cursor_promote_node16_to_node48() {
                 MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap()
+                    &TrieHash::from_bytes(&path[..]).unwrap()
                 )
                 .unwrap()
                 .unwrap(),
@@ -734,7 +734,7 @@ fn trie_cursor_promote_node16_to_node48() {
                 path[k] = j + 40;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -755,7 +755,7 @@ fn trie_cursor_promote_node16_to_node48() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -784,7 +784,7 @@ fn trie_cursor_promote_node16_to_node48() {
             path[k] = 129;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -806,7 +806,7 @@ fn trie_cursor_promote_node16_to_node48() {
                 MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap()
+                    &TrieHash::from_bytes(&path[..]).unwrap()
                 )
                 .unwrap()
                 .unwrap(),
@@ -897,7 +897,7 @@ fn trie_cursor_promote_node48_to_node256() {
                 path[k] = j + 32;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -918,7 +918,7 @@ fn trie_cursor_promote_node48_to_node256() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -947,7 +947,7 @@ fn trie_cursor_promote_node48_to_node256() {
             path[k] = 128;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -968,7 +968,7 @@ fn trie_cursor_promote_node48_to_node256() {
                 MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap()
+                    &TrieHash::from_bytes(&path[..]).unwrap()
                 )
                 .unwrap()
                 .unwrap(),
@@ -1004,7 +1004,7 @@ fn trie_cursor_promote_node48_to_node256() {
                 path[k] = j + 40;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -1024,7 +1024,7 @@ fn trie_cursor_promote_node48_to_node256() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -1053,7 +1053,7 @@ fn trie_cursor_promote_node48_to_node256() {
             path[k] = 129;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -1074,7 +1074,7 @@ fn trie_cursor_promote_node48_to_node256() {
                 MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap()
+                    &TrieHash::from_bytes(&path[..]).unwrap()
                 )
                 .unwrap()
                 .unwrap(),
@@ -1110,7 +1110,7 @@ fn trie_cursor_promote_node48_to_node256() {
                 path[k] = j + 90;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -1131,7 +1131,7 @@ fn trie_cursor_promote_node48_to_node256() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -1160,7 +1160,7 @@ fn trie_cursor_promote_node48_to_node256() {
             path[k] = 130;
 
             let mut c =
-                TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
             let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
 
@@ -1181,7 +1181,7 @@ fn trie_cursor_promote_node48_to_node256() {
                 MARF::get_path(
                     &mut f,
                     &block_header,
-                    &TriePath::from_bytes(&path[..]).unwrap()
+                    &TrieHash::from_bytes(&path[..]).unwrap()
                 )
                 .unwrap()
                 .unwrap(),
@@ -1256,7 +1256,7 @@ fn trie_cursor_splice_leaf_4() {
                 path[5 * k + 2] = 32;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 test_debug!("Start splice-insert at {:?}", &c);
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
@@ -1283,7 +1283,7 @@ fn trie_cursor_splice_leaf_4() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -1349,7 +1349,7 @@ fn trie_cursor_splice_leaf_2() {
                 path[3 * k + 1] = 32;
 
                 let mut c =
-                    TrieCursor::new(&TriePath::from_bytes(&path[..]).unwrap(), f.root_trieptr());
+                    TrieCursor::new(&TrieHash::from_bytes(&path[..]).unwrap(), f.root_trieptr());
 
                 test_debug!("Start splice-insert at {:?}", &c);
                 let (nodeptr, mut node, node_hash) = walk_to_insertion_point(&mut f, &mut c);
@@ -1372,7 +1372,7 @@ fn trie_cursor_splice_leaf_2() {
                     MARF::get_path(
                         &mut f,
                         &block_header,
-                        &TriePath::from_bytes(&path[..]).unwrap()
+                        &TrieHash::from_bytes(&path[..]).unwrap()
                     )
                     .unwrap()
                     .unwrap(),
@@ -1413,7 +1413,7 @@ where
         for i in 0..count {
             eprintln!("{}", i);
             let path = path_gen(i);
-            let triepath = TriePath::from_bytes(&path).unwrap();
+            let triepath = TrieHash::from_bytes(&path).unwrap();
             let value = TrieLeaf::new(
                 &vec![],
                 &[
@@ -1519,7 +1519,7 @@ where
 
         for i in 0..count {
             let path = path_gen(i);
-            let triepath = TriePath::from_bytes(&path).unwrap();
+            let triepath = TrieHash::from_bytes(&path).unwrap();
             let value =
                 MARF::get_path(&mut marf.borrow_storage_backend(), &block_header, &triepath)
                     .unwrap()